% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dominance.R
\name{find_dominance}
\alias{find_dominance}
\title{Calculate Dominance Index with Normalization}
\usage{
find_dominance(theta, doc_indices, n = 3)
}
\description{
Calculate Dominance Index with Normalization
}
